begynt å refaktorere fra å ha all logikk i mapValues
